+++ CHANGELOG.md
@@ -14,6 +14,13 @@ Versioning](https://semver.org/spec/v2.0.0.html).
   loading the library with `RTLD_NOW`. This test is only run on Unix-like
   platforms.
 
+### Changed
+
+- When a plugin supports text-to-value and/or value-to-text conversions for some
+  but not all of its parameters, clap-validator will now include the names of
+  the parameters and the failing inputs in the error message to help pinpoint
+  the issue.
+
 ## [0.2.0] - 2022-01-09
 
 ### Added